requirements.txt
streamlit~=1.40.2
python-docx~=1.1.2
protobuf~=5.29.1
ollama~=0.4.4
rich~=13.9.4
google-generativeai